spelling_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pillow==10.1.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2 